Simple implementation of DNSMasq backend
The logic is quite simple. All dns zones are stored in separate files.
They are eventually merged into one flat dns zone.
And this file (or any separate zone) should be specified in dnsmasq config file (--addn-hosts=).
Update is also simple, all you need is to send a SIGHUP to dnsmasq to force re-read of the file.

Change-Id: Id634ca2843fd02a8406913881d4989c7c2112249