Update domains when servers are created, modified or deleted
Existing domains do not get updated in the backend when new servers are
added or existing ones modified or deleted.
This is rectified in this fix by implementing methods in all of the backends.
Three abstract methods were added to the base module that all backends
inherit from and implemented in all backends.
The three methods implemented are create_server, update_server, and
delete_server.
The methods are called from central/service.py as needed.

Change-Id: I1ac9add27f91ee6d73fdcfda5b2ef8c0c1ecb0ab
Closes-Bug: #1214869