Add SLDs that act like TLDs
This change offers the designate administrator the ability to configure
2 lists to indicate the TLDs(accepted_tlds_file) and SLDs and others
that
effectively act as TLDs.(effective_tlds_file)

2 sample files are included to show the format required for these files.
The TLD file is from http://data.iana.org/TLD/tlds-alpha-by-domain.txt
Commented lines in this file begin with a #. There is one entry per line
and the entries are in the IDNA format.

The sample effective TLD file is from http://publicsuffix.org/list/
This file is in the UTF-8 format.  Lines beginning with a // and ! are
ignored.  The wildcard character * (asterisk) may only be used to
wildcard
the topmost level in a domain name.  All the entries  are converted to
idna format.

Closes-Bug: #1249396

Change-Id: I9718489335e24cb24733d4c4fc2a966e490014d3