Add SLDs that act like TLDs
This change is from the Icehouse release. The effective TLD file need
not be present. When it is not present the behavior would be the same
as before. When it is present, the domains that are being created are
checked to see that it is not an effective TLD.

This change offers the designate administrator the ability to configure
2 lists to indicate the TLDs(accepted_tlds_file) and SLDs and others
that effectively act as TLDs.(effective_tlds_file)

2 sample files are included to show the format required for these files.
The TLD file is from http://data.iana.org/TLD/tlds-alpha-by-domain.txt
Commented lines in this file begin with a #. There is one entry per line
and the entries are in the IDNA format.

The sample effective TLD file is from http://publicsuffix.org/list/
This file is in the UTF-8 format. Lines beginning with a // and ! are
ignored. The wildcard character * (asterisk) may only be used to
wildcard the topmost level in a domain name. All the entries are
converted to idna format.

Change-Id: I3d2f7e2741f4cda0b227e4a5671b141f99052291
Closes-Bug: #1249396