Make sure that certfile and keyfile are readable
If the certfile or keyfile is missing or Designate doesn't have
permission to read them, nsd4slave would throw a cryptic SSL exception
when it tried to connect to the NSD4 servers. To avoid this, make sure
that the certfile and keyfile are readable upon backend init and raise a
human-readable exception if they're not.

Change-Id: I837cd328cf31f568e5da4ad4edf8c4dd34ddf8e2