Add APIs for managing TLDs
This change adds APIs for managing TLDs. Previously TLDs were in
2 files. With this change, the TLDs will be in the central's storage
database.

When the service starts for the first time, by default there are no
TLDs in the database.  When there are no TLDs in the database
the behavior is to not check for TLDs when creating domains.
Once a TLD is created in the database, the TLD checks are
enforced.

Change-Id: Ibc945af3b158355475f5c8ee7887ed4b32081337