Import TLDs to Designate
Currently when Designate starts, it does not have any TLDs. TLDs need
to be added using the API. This change provides an easier way to
add/import/migrate TLDs, rather than calling the API repeatedly.

The TLDs need to be provided in a csv file. Each line in this file
contains a TLD entry followed by an optional description.

If any lines in the input file result in an error, the program
continues to the next line.

Change-Id: I6f5abbdbbebb27f7f238c2ccb99fe5f9b2aee6c4
Implements: blueprint import-tlds