Default state-path to /var/lib/designate
In a production system, the previous state-path would default to
something like /usr/lib/python2.6/site-packages. Not only is this a
bizarre place to keep Designate's state, it also breaks the bind9
backend, which attemps to create and write to
/usr/lib/python2.6/site-packages/bind9.

Since /var/lib is defined in the Filesystem Hierarchy Standard as the
place to store persistent state information, it's better to explicitly
use it as the default state-path.

Change-Id: I91922799ba1dcae1ad33f1abbec79b1eb907307f
Closes-bug: 1279989