Ensure that request body in v2 contains valid fields
Currently 'id' and other fields like 'created_at' and 'updated_at' can
be set in the request body.  Some of these fields are ignored and
some like 'id' can actual set the id.
With this change - remove 'id' from the list of valid keys and ensure
that only valid keys are present in the request body.
Closes-Bug: #1291518

Change-Id: I99f3f106c0004fa5521869e46da85f1053241eb5