Post oslo.messaging fixes
If rpc isn't inited before attempting to load the backend backends relying on
rpc will fail. For Central this should happen at the start of the __init__.

Also the agent_api should be a attribute on the backend object rather then a
module variable.

Pass the domain in the create_domain call in the rpc backend.

Change-Id: I85445963b091d1fbed33c132ae372e3f90cb0c68