add sink support to designate devstack
Add support for designate-sink to the designate devstack plugin.
There can be a different domain used for fixed IPs and floating
IPs.  By default, the nova_fixed handler will be used for fixed
IPs, and the neutron_floatingip handler will be used for floating
IPs.  The domains will be created if they do not already exist.
In order to use the nova_fixed handler, nova must be enabled,
and DESIGNATE_TEST_DOMAIN_FIX must be defined.
In order to use the neutron_floatingip handler, neutron must be enabled,
and DESIGNATE_TEST_DOMAIN_FLT must be defined.
This will also configure nova and neutron to emit notifications if they
are not already configured to do so.

Change-Id: Ib3642212e0fdf42de2cbdf8289fb6465c935a95f
Closes-Bug: #1315187