Fix FreeIPA tests mock's
Bug 1321873 was triggered by the addition of the FreeIPA tests,
where they mocked out two of the Central services methods, but
failed to restore the original methods after each test completed.

The mocked methods returned very little data compared to the real
methods, resulting in the KeyError. The nondeterministic way in
which testr executes our tests results in this failure only
appearing on 1 of N test runs.

Change-Id: Id9428fb9c344ea547398a6a15f7af78d12fa7221
Closes-Bug: 1321873