Setup for translation
To start translation, we need to initially import the
translation file - and place it at the proper place so that
the usual CI scripts can handle it.

The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg.

Add also babel.cfg file as translation configuration file.

Further imports of po files will be done by the OpenStack Proposal bot.

Change-Id: I679c3da7aca1911f21ddc8551a352afdbf3a4f78