Avoid logging "Expected" exceptions returned over RPC
When an expected exception, such as a NotFound or Duplicate is passed over
RPC, we identify it via the `expected` attribute and raise oslo.messagings
ExpectedException. The internals of ExpectedException effectively save the
current exception details (type, trace, etc) inside the ExpectedException,
when received on the client side, the ExpectedException is detected, and
the saved details are used to bubble the original exception back up. On
both sides of the RPC, logging is avoided when ExpectedException is used.

Change-Id: I143882181e3346c4ebe1f0ad491c75a4db549fe4
Closes-Bug: #1304152