Migrate to SQLAlchemy Core
SQLAlchemy ORM is providing no advantages to us, while at the same time
contributing towards the difficutly in implementing complex queries. This
is *not* a SQLAlchemy ORM issue, it's simply our use of the ORM.

Rather than fixing our use of the ORM to be more intelligent, we should
simply drop down to SQLA Core, as the ORM provides no direct benefits.

The PowerDNS backend should also be updated to match, as a separate
review.

Change-Id: Ife5f5740821a23ff119c27c1c22cba3eca837f1b