Fix warnings made by keystone_authtoken
auth_* config is deprecated in keystone_authtoken. This change
fixes the log "Configuring admin URI using auth fragments. This
is deprecated, use 'identity_uri' instead." in the gate logs.

The patch also configures auth_uri because keystone_authtoken now
requires it to be set (see the gate log "Configuring auth_uri to point to the
public identity endpoint is required; clients may not be able to
authenticate against an admin endpoint")

Change-Id: I5e067a3e4900e7eb5e0a1230532834e43538145a