Return a status field from the recordsets API
Currently on a get for a recordset, there is no status
returned. In the database schema, status only exists on
records.

The Recordset object now has 2 read only fields - action
and status. The action and status fields are calculated
based on the  records. The status is the worst of all the
records in order of ERROR, PENDING, ACTIVE.

action is set to UPDATE if present. CREATE and
DELETE are set if they are the only ones. Otherwise
action is set to NONE

If there are no records in the recordset the action and
status would always be 'NONE' and 'ACTIVE'.

Change-Id: Idca4d30db67e2bb29d08e11b3d22685422a6b5e9
Closes-Bug: 1409824