Assume a new developer won't create a "designate" user in the dev guide
The getting started docs suggest checking the code out into `/var/lib`,
but this requires a later step of adding a `designate` user to the
sudoers list without adding a step of adding the `designate` user in the
first place.

This updates the docs and basic config example to simply check out the
code into `$HOME/openstack/designate/` and removes sudoers step.

Closes-Bug: 1423628
Change-Id: Ia30b5460532589417a7e8b1fad9c233ced954106