Add error state transition for creation of domains
* If a domain creation is not successful on the backend
servers, then the domain transitions to an error state.
* Rewrote _clear_cache to skip an explicit retrieve.
For memcache no retrieve is necessary. So no additional
action is needed.
For sqlalchemy, an id is needed for a delete, so a retrieve
is done on a clear if needed.
* periodic_sync now sync's domains that are not in an
error state (periodic_recovery handles domains that are
in error)
* Added/modified unit tests

Change-Id: I375262b3ccf4d1d9da1c8954f4768a55955a0129
Closes-Bug: 1413800
Closes-Bug: 1425695