Service Class Cleanup - Part 2/3
Here, we rework the various Service base classes to be mixins rather than
subclasses of the base Service class, in preperation for allowing for MiniDNS
to be both a RPC and DNS Service - sharing code with Central etc (RPCService)
and the Agent (DNSService).

Part three will move MiniDNS to using the DNSService mixin.

Change-Id: I417957026705bd4511af60bcbec4852970e82fcb