Service Class Cleanup - Part 3/3
Here, we rework the MiniDNS service to use the DNSService class, and move
code that was temporarily placed in the dnsutils module into the DNSService
class. Once this is merged, the newly moved code will be updated to handle
TSIG verification for the mDNS service.

Long term, we should probably follow the RPCService vs RPCServer (from
oslo.messaging) model, allowing for getting all of the packet parsing
code into a single separate class.

Change-Id: I2ef936e570a8c19c7b0145a32e0aed1ab0718fa7