Fix handling of wildcards in api / storage
We need to move the handling of '*' to '%' up to the API layer since if not it
causes a issue described in the bug. This change addresses this by removing the
replace of '*' to '%' in the storage layer to the API and also fixes any other
places '*' is used for wildcard search.

Closes-Bug: #1391426

Change-Id: Ib75e8359b8040effea3dfdbf9c2baf05c3ed208b