DevStack: Akamai/Dyn Support for Jenkins Credentials and Exercises
The Jenkins credential binding plugin will provide the secrets in
a single colon separated environment variable, so we pass this
through and update the Akamai and Dyn drivers to support reading
from it.

The exercices are also now fixed, where both the Akami and Dyn
backends would result in more than 1 NS records, the original
counters were off.

Change-Id: I7bb37e9671dae34b47abf9bf092d90e8ba248ccb