Correct pool_target options in sample config
When https://review.openstack.org/#/c/170612/ landed, the host/port
configuration for pool targets moved to be regular pool_target options.

The config that is shown in the documentation reflects this, but the sample
config does not. This gets really confusing, because the default host/port
is 127.0.0.1:53, so the Pool Manager will still start, and NOTIFYs will
be sent there.

Change-Id: If682bdcdf2bfb5f2a6c7f435a41e0577e262a5f3