Added CORS support to Designate
This adds the CORS support middleware to Designate, allowing a deployer
to optionally configure rules under which a javascript client may
break the single-origin policy and access the API directly.

The paste.ini method of deploying the middleware was
chosen, because it needs to be able to annotate error responses
created by other middlewares. If one such middleware throws an
error - such as an auth error from keystone - that error response
must still have CORS headers in order to be readable by the
user agent.

openstack-config-generator hooks were not added in this patch, as
it does not appear that designate explicitly supports them.

OpenStack CrossProject Spec:
   http://specs.openstack.org/openstack/openstack-specs/specs/cors-support.html
Oslo_Middleware Docs:
   http://docs.openstack.org/developer/oslo.middleware/cors.html
OpenStack Cloud Admin Guide:
   http://docs.openstack.org/admin-guide-cloud/cross_project_cors.html

Change-Id: Ib66e8afafbde568549228c98d210f02466a90cc8