Fix argument order in assertEqual to (expect, obs)
assertEqual expects that the arguments provided to it should be
(expected, observed). If a particular order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit Tests
fail. There are several Unit Test files where the arguments for assertEqual
have been swapped.

Change-Id: I6fdedcde67f76883257d35d1b812ab8a0e3e0733
Related-Bug: #1259292