Remove py26 support from designate
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I4aeba7269ddb465ac61f851f5b5d2ab1b91882e4
Related-Bug: #1519510