Use lazy properties in services
Seeing as it is common to mock / patch the storage (and any service
interface point), this makes the quota and storage attributes lazy
properties that will only be loaded upon usage. This allows
instantiating the class without a database connection as well as easy
patching via setting an attribute.

For example:

    def test_create_zone(self):
        central = Service()
	central._storage = Mock()
	central.create_zone(self.context, self.zone)
	...

Change-Id: I189559738a5a43001bbc4bdfe94659ad8c58143b