Add noop metrics client
With the addition of monasca-statsd as part of
I990472fa059afde37f7e4a1284360c17162aab49, it's a required dependency
even if not used.  For upstream packages which do not have the
monasca-statsd available, we need to be able to continue to have
designate working without it. This change makes the monasca-statsd
dependency optional at runtime by falling back to a noop client that
mocks out the interface for monasca-statsd but does nothing.

Change-Id: Idf986c514b26f1d2799903b2a4e581a1166343cc
Closes-Bug: #1630666