Use tenacity instead of retrying
We are replacing all usages of the 'retrying' package with 'tenacity'
as the author of retrying is not actively maintaining the project.
Tenacity is a fork of retrying, but has improved the interface and
extensibility (see [1] for more details). Our end goal here is removing
the retrying package from our requirements.
[1] https://github.com/jd/tenacity

Co-Authored-By: ShuYingya<yingya.shu@easystack.cn>
Change-Id: I68b8f7abab0ca8ab3c15b346aebea261cc26486b
Closes-Bug: #1635384