Fix noop client getters
During additional testing, we found that we needed to actually return an
object for the getters or an attribute error is thrown for the timed
property.  This change fixes the getters to properly return the object
and adds unit tests for both the noop and monasca clients to ensure the
right objects are returned for the metrics class. Additionally, gauge is
not spelled guage.

Change-Id: I2b229069f925306b1e48b92f2f3eacb2ded39aa6
Closes-Bug: #1639248