Fix servicing IPv6 based TCP connections
When a connection is received via IPv6, the address is a 4-tuple
instead of a 2-tuple for IPv4, so handle this accordingly.

Also handle the end of the TCP connection gracefully instead of
emitting an "Invalid packet" warning every time.

[1] https://docs.python.org/2/library/socket.html

Change-Id: I935bbea1d58af8a5b41b5d29f12de4dcb7cf6301
Closes-Bug: 1734101
(cherry picked from commit 5e45b15419e80539e35d08380351445553bf92b0)