Resolve get_dns_view error after selecting multi-tenant view
Adds a second comparison check when deciding whether to display the single
tenant view or multi-tenant view in Infoblox. The issue is that the DB field
is set to '0', which is not interpreted as false.

Change-Id: Id0fac71cc6d82123a7ee5d5983467acd1b8e5a9e
Closes-Bug: 1752349
(cherry picked from commit e4de365775c1f6f2722f2ad3109f3349d9a5a691)