Use openstack-single-node-bionic nodeset
designate job designate-bind9-py36 use the nodeset
name 'openstack-bionic-node' which is defined in Tempest.

Tempest is plannig to remove that nodeset which is duplicate of
what devstack has - https://review.openstack.org/#/c/641735/

This job is only running in Rocky branch so changing it directly
in rocky.

This commit modifies the job to use the openstack-single-node-bionic
which is latest defined in devstack.

Change-Id: Id24b813dc4515af130682d12bada075ecf4b277c