Ensure pdns4 driver is consistent with other drivers
Other drivers will delete and / or overwrite zones that pre-exist on a server.
Updates the pdns4 driver to act the same way, and also clean up zones
that were not created properly.

Also send a notify immedidiately after creation to force a sync from the pdns
server, instead of waiting for the timer to trigger the first transfer.

Change-Id: I59fe82220adef4b28d4621e093064ee971e1dea8
Signed-off-by: Graham Hayes <gr@ham.ie>