Only log unexpected rndc errors
A common false positive when bind is failing
to create a zone is the output that the zone
already exists.

This is usually due to a misconfiguration,
and not because the rndc command is failing.

We still log the rndc output using debug,
and log it in cases where it only would
fail due to legit issues talking to bind.

Change-Id: If0f6dd84959f67f32394084ea66d54e210dad966