Fixed neutron endpoint override
The network_api endpoint code would never
actually use the manually configured endpoints.

This patch changes that by first checking if
custom endpoints were configured, and if that
isn't the case it will use the service_catalog
instead.

In addition we introduce full testing on
this code path to make sure it behaves
as we expect.

Change-Id: I2e9f1485fce401336ab8a4a8b1aa1f971292168f