Fix designate-sink shutdown issue
If no notification handler is configured the
designate-sink may not shutdown properly.
This is because the listener was created but never
started. This caused oslo.messaging to hang on
waiting for a successful start before stopping.

Closes-Bug: #1877201
Change-Id: Ie51f7c3d64cbb48ae359256b6fa5a0196fb3f248
(cherry picked from commit 40cbc9bd43d3b43d6cdc4daeb5a089df96603ed7)