Fix issues with newer flake8
The updated flake8 seems to have fixed a check for .format() issues that
discovered a bug in one of our tests, which is fixed here.

Also get rid of using "l" as a variable name and tag the imports that
need to stay in order after setting an environment variable.

Change-Id: I28de0ffa68f9bff4685d81fe7b5a5d8909e2313c