Make greande jobs n-v for EM and oldest stable
As discussed in ML thread[1], we are going to
make grenade jobs as non voting for all EM stable and
oldest stable. grenade jobs are failing not and it might take
time to fix those if we are able to fix. Once it jobs are
working depends on project team, they can bring them back to
voting or keep non-voting.

If those jobs are failing consistently and no one is fixing them
then removing those n-v jobs in future also fine.

Also pin designate-tempest-plugin to the latest tag that still supports
python2.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-June/015499.html

StableOnly

Change-Id: I8eb1a244ac9d79c9915c222af5afd298bdb04997