New Doc intro section
Added a new section to the documentation introducing the core
concepts of DNS, providing an overview of Designate and how
it integrates with Neutron and external nameservers and linking
to other documentation for more information.

The Designate architecture diagram has been updated to reflect
modern deployments:
  - All services are run as HA
  - Nova does not interact with the Designate API, it has been
    replaced with 'Users'
  - The DB only receives connections from MiniDNS or Central so
    its arrows are all incoming
  - The backend is a part of the worker service
  - MiniDNS sends NOTIFY to customer DNS Servers and
    also receives transfer requests so their connection is
    bidirectional

A subsequent change can update the architecture section to reflect
these clarifications, though they are for the most part already
mentioned in the text there.

Change-Id: I471db98544332cb454d15f29d86407cd09e91d6c