Implement secure RBAC for tenant policies
This commit updates the tenant policies to understand scope checking
and account for a read-only role. This is part of a broader series of
changes across OpenStack to provide a consistent RBAC experience and
improve security.

I'm not entirely sure I understand these policies. It'll be good to
work through these policy changes with someone more familiar with
desginate and why these policies exist.

Change-Id: I9b6bce0c43720f61cdebfa416d953e5a2b920e87