Revert "Move back to mod_wsgi for devstack"
This reverts commit 185d92d230bdf8149b3ae3f6d1730ec23c0514d7.

Reason for revert: Lets try to move back to uwsgi, as logging is a lot cleaner in the CI with uwsgi. Switches uwsgi from using uwsgi over socket to over localhost HTTP.

This patch also bumps the API_WORKERS setting for devstack from 2 to 4.
This will bump the uwsgi processes and threads from two to four.

Change-Id: I126133564e2ada32448ad36e0d021790939425d4