Fix grenade upgrade API check URL path
Previously, the grenade API started check during the upgrade was
checking if the root of the web server was functioning. In this job
that is testing the horizon endpoint and not the designate API.
This patch fixes this to have the check run against the designate
endpoint under apache. This prevents false failures when horizon
has an issue.

Change-Id: Idbfec1adca2024cd5f352017a7c9319dcec65d42