Fix update blacklist policy deprecation
The update blacklist policy deprecation was using the wrong variable and
referenced the policy for get blacklist. This was a typo and causes
oslo.policy to alias the old name with the new name since it things the
policy name is changing, which isn't the case.

This commit updates the policy to use the correct deprecated policy
rule.

Change-Id: I702a4cbef6d580249d17875af34b35822c0c9d32