Allow TXT record over 255 characters if split
As defined in RFC1035 section 3.3.14 TXT-DATA
can be one or more <character-strings>s.
Before this commit Designate threw errors
when saving TXT records that are split
into multiple strings because validations on
the field did not consider allowing this
in a single DNS record as per RFC7208
section 3.3.

This patch allows longer TXT record data
but only if it is split according to RFC
definitions mentioned above. If data
is made of more <character-string>s,
each one is individually validated with
the same validations as if the data was
not split.

Closes-Bug: 1595265
Change-Id: I4e3e51b32ab01efc4202c297708eff5a2e2b4985
Signed-off-by: Emanuel Andrecut (emanuel.andrecut@fleio.com)