Fix designate-manage pool update bugs
This patch addresses a few problems with the manage
command for pools and NS records.

- Fixed an issue where having multiple NS records
  would break the pool command.

- Fixed a scenario where manually (non-managed) NS
  records could break the pool command.

- Fixed a potential edge case that could break the
  pool command.

The biggest change is that we now only manage the NS
record for the zone itself. This was always the case,
but because we didn't check for this specifically,
other NS records would conflict with the command.

Change-Id: I4e6ea0b6b717d2a1b5cc420874d1bb8fb290e04b
(cherry picked from commit bff3d5f6e31fe595a77143ec4ac779c187bf72a8)