Fix duplicate zone when creating ptr records
This fixes a race-condition when creating multiple PTR records
under the same zone. There is a brief window when creating two
identical zones can cause an error. This adds a fallback that
should prevent the error.

I also added a threaded test that caught multiple additional
bugs in this code.
- Wrong find_recordset used caused the wrong exception to be thrown.
- The transaction workflow would break error handling.

Change-Id: Ia1194ab838c52d5d91cb1d26c4556c73b4f3a745